+++ demosplan/DemosPlanCoreBundle/CustomField/RadioButtonField.php
@@ -12,10 +12,7 @@ namespace demosplan\DemosPlanCoreBundle\CustomField;
 
 class RadioButtonField extends AbstractCustomField
 {
-
     protected string $id = '';
-
-    protected string $name = '';
 
     protected string $fieldType = 'singleSelect';
 
@@ -89,14 +86,17 @@ class RadioButtonField extends AbstractCustomField
         $this->fieldType = $type;
     }
 
-
     public function getType(): string
     {
         return 'singleSelect';
     }
 
-    public function isValueValid(string $value): bool
+    public function isValueValid(?string $value): bool
     {
+        if (null === $value) {
+            return true;
+        }
+
         if (in_array($value, $this->options, true)) {
             return true;
         }