+++ demosplan/DemosPlanCoreBundle/Repository/CustomFieldConfigurationRepository.php
@@ -42,12 +42,17 @@ class CustomFieldConfigurationRepository extends CoreRepository
     public function findCustomFieldConfigurationByCriteria(
         string $sourceEntity,
         string $sourceEntityId,
-        string $targetEntity
+        string $targetEntity,
+        ?string $customFieldId = null,
     ): ?array {
         try {
             $criteria = ['sourceEntityId' => $sourceEntityId];
             $criteria['sourceEntityClass'] = $sourceEntity;
             $criteria['targetEntityClass'] = $targetEntity;
+
+            if ($customFieldId) {
+                $criteria['id'] = $customFieldId;
+            }
 
             return $this->findBy($criteria);
         } catch (Exception $e) {
@@ -61,7 +66,7 @@ class CustomFieldConfigurationRepository extends CoreRepository
         string $sourceEntity,
         string $sourceEntityId,
         string $targetEntity,
-        CustomFieldInterface $customField
+        CustomFieldInterface $customField,
     ): CustomFieldConfiguration {
         $customFieldConfiguration = new CustomFieldConfiguration();
 
@@ -105,7 +110,6 @@ class CustomFieldConfigurationRepository extends CoreRepository
             $newCustomFieldConfiguration->setConfiguration($customFieldConfiguration->getConfiguration());
             $this->add($newCustomFieldConfiguration);
         }
-
     }
 
     public function getCustomFields(string $sourceEntity, string $sourceEntityId, string $targetEntity): ArrayCollection
@@ -121,6 +125,7 @@ class CustomFieldConfigurationRepository extends CoreRepository
                 static function (CustomFieldConfiguration $customFieldConfiguration): CustomFieldInterface {
                     $customField = $customFieldConfiguration->getConfiguration();
                     $customField->setId($customFieldConfiguration->getId());
+
                     return $customField;
                 },
                 $customFieldConfigurations