+++ CHANGELOG.md
@@ -5,6 +5,31 @@
 - **Patch Version**: Incremented for bug fixes.
 
 ## UNRELEASED
+- Allow to configure procedures to accept or not anonymous statements
+- Allow filtering of institution tags in AdminstrationMemberList / refactor twig
+- Add configurable feedback control for public participation statements
+- Migrate to Tailwind CSS v4
+
+
+## v4.5.0 (2025-06-25)
+- Export Original Statements as docx in the Statement List
+- Allow filtering of institution tags in DpAddOrganizationList
+
+## v4.4.0 (2025-06-13)
+## v4.3.0 (2025-06-13)
+- Add Versioning of custom fields of segments
+- Update Elasticsearch to version 8
+- Export Original Statements as csv in the Statement List
+
+## v4.1.0 (2025-05-21)
+## v4.0.0 (2025-05-21)
+- Update to symfony 6.4
+
+## v3.3.0 (2025-05-13)
+- restore deleted logger entry 
+
+## v3.2.0 (2025-05-13)
+- Enable Custom Field feature on segments: Allow users to add/edit custom fields to their segments     based on the custom fields defined in the procedure
 - Add Custom Field feature: Allow users to add custom fields to their procedures
 
 ## v2.27.1 (2025-04-09)